requirements.txt
Flask==3.1.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.46
alembic==1.18.3
PyMySQL==1.1.2
python-dotenv==1.2.1
PyJWT>=2.8.0
gunicorn>=21.0.0
